// src/services/obsidianRestAPI/generated-types.ts
export interface paths {
// ...
  "/search/simple/": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    put?: never;
    /** Search for documents matching a specified text query
     *      */
    post: {
      parameters: {
        query: {
          /** @description Your search query */
          query: string;
          /** @description How much context to return around the matching string */
          contextLength?: number;
        };
        header?: never;
        path?: never;
        cookie?: never;
      };
      requestBody?: never;
      responses: {
        /** @description Success */
        200: {
          headers: {
            [name: string]: unknown;
          };
          content: {
            "application/json": {
              /** @description Path to the matching file */
              filename?: string;
              matches?: {
                context: string;
                match: {
                  end: number;
                  start: number;
                };
              }[];
              score?: number;
            }[];
          };
        };
      };
    };
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
  schemas: {
    Error: {
      /**
       * @description A 5-digit error code uniquely identifying this particular type of error.
       *
       * @example 40149
       */
      errorCode?: number;
      /**
       * @description Message describing the error.
       * @example A brief description of the error.
       */
      message?: string;
    };
    NoteJson: {
      content: string;
      frontmatter: {
        [key: string]: unknown;
      };
      path: string;
      stat: {
        ctime: number;
        mtime: number;
        size: number;
      };
      tags: string[];
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}